% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{download_exuber}
\alias{download_exuber}
\title{Download nternational House Price Database Exuberance Data}
\usage{
download_exuber(option = c("gsadf", "bsadf"))
}
\arguments{
\item{option}{which to download}
}
\description{
Download nternational House Price Database Exuberance Data
}
