% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{download_price}
\alias{download_price}
\title{Download nternational House Price Database Data}
\usage{
download_price()
}
\description{
All releases of the database include the following time series for the
countries covered:
\itemize{
\item The house price index (hpi).
\item The house price index expressed in real terms (rhpi).
\item The personal disposable income index (pdi).
\item The personal disposable income expressed in real terms index (rpdi).
}
All variables in real terms are deflated with the personal consumption
expenditure (PCE) deflator.
}
