% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ihpd_countries}
\alias{ihpd_countries}
\title{List all the countries}
\usage{
ihpd_countries(version = NULL)
}
\arguments{
\item{version}{Which version to download. Version number should be a character
of the following format \%Y\%q (e.g. '1801' - corresponds to year 2018,
Quarter 1). Versions start from '1102'. Defaults at the latest available.}
}
\value{
A character vector all the countries.
}
\description{
List all the countries
}
\examples{
ihpd_countries()

}
