% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ihpd_release_dates}
\alias{ihpd_release_dates}
\title{Fetches the latest release dates}
\usage{
ihpd_release_dates()
}
\value{
A data.frame with 2 varibles with the latest release dates data.
}
\description{
Fetches the latest release dates
}
\examples{
ihpd_release_dates()
}
