% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{ihpd_browse}
\alias{ihpd_browse}
\title{Quickly browse to important links}
\usage{
ihpd_browse(pick = c("info", "app"))
}
\arguments{
\item{pick}{whether to browse the shiny \code{info} or the \code{app} page. Defaults at
\code{info}.}
}
\description{
This function take you to webpages associated with the International House
Price Database and returns the url invisibly. Option \code{app} take you the
International Housing Observatory, while \code{info} take you to Dallas Fed
info page of the data.
}
\examples{

ihpd_browse()
}
