% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ihpd_get}
\alias{ihpd_get}
\title{Download Data from the International House Price Database}
\usage{
ihpd_get(symbol = c("raw", "gsadf", "bsadf"), version = NULL)
}
\arguments{
\item{symbol}{Which dataset to download.}

\item{version}{Which version to download. Version number should be a character
of the following format \%Y\%q (e.g. '1801' - corresponds to year 2018,
Quarter 1). Versions start from '1102'. Defaults at the latest available.}
}
\value{
Returns a tibble in long format.
}
\description{
Available downloads:
\enumerate{
\item raw: Raw Data
\item gsadf: Generalized sup ADF statistics
\item bsadf: Backward sup ADF statistic sequence
}
}
\details{
\describe{
\item{raw}{

Includes the following time series for the countries covered:
\itemize{
\item hpi: The house price index.
\item rhpi: The house price index expressed in real terms.
\item pdi: The personal disposable income index.
\item rpdi: The personal disposable income expressed in real terms index.
}
All variables in real terms are deflated with the personal consumption
expenditure (PCE) deflator.}
}

The exuberance indicators include
\describe{
\item{gsadf}{
\itemize{
\item type: real house price index (rhpi) and the ratio of rhpi to real
personal disposable income (rpdi)
\item lag: lag length of 1 and 4
\item sig: significance level for the critical value
\item value: test statistics for explosive behavior
\item crit: 95 percent critical values.
}
}

\item{bsadf}{
The exuberance indicators include
\itemize{
\item type: real house price index (rhpi) and the ratio of rhpi to real
personal disposable income (rpdi)
\item lag: lag length of 1 and 4,
\item value: test statistics for explosive behavior
\item crit: 95 percent critical values.
}
}
}
}
\examples{

ihpd_get()
}
